[#5] Refactoring
