+++ src/test/java/cz/sio2/obo/versionextractor/TestOWLVersionExtractor.java
@@ -1,7 +1,8 @@
-package cz.sio2.obo;
+package cz.sio2.obo.versionextractor;
 
+import cz.sio2.obo.Version;
+import cz.sio2.obo.versionextractor.FSVersionExtractor;
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvFileSource;
 
@@ -11,18 +12,18 @@ import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.Objects;
 
-public class TestXMLVersionExtractor {
+public class TestOWLVersionExtractor {
 
     @ParameterizedTest
-    @CsvFileSource(resources = "/xml-testcases.csv", numLinesToSkip = 1, delimiter = ',')
+    @CsvFileSource(resources = "/owl-testcases.csv", numLinesToSkip = 1, delimiter = ',')
     public void testExtractVersionIriFromXMLExtractsVersionInfoCorrectlyIfPresent(
             String file,
             String ontologyIri,
             String versionIri,
             String versionInfo) throws URISyntaxException, IOException {
-        final String s = Files.readString(Paths.get(Objects.requireNonNull(getClass().getResource("/xml-testcases/" + file)).toURI()));
+        final String s = Files.readString(Paths.get(Objects.requireNonNull(getClass().getResource("/owl-testcases/" + file)).toURI()));
         final Version version = new Version();
-        new XMLVersionExtractor().extract(s, version);
+        new FSVersionExtractor().extract(s, version);
         Assertions.assertEquals(ontologyIri, version.getOwlOntologyIri());
         Assertions.assertEquals(versionIri, version.getOwlVersionIri());
         Assertions.assertEquals(versionInfo, version.getOwlVersionInfo());
