+++ src/test/java/cz/sio2/obo/versionextractor/TestXMLVersionExtractor.java
@@ -1,5 +1,7 @@
-package cz.sio2.obo;
+package cz.sio2.obo.versionextractor;
 
+import cz.sio2.obo.Version;
+import cz.sio2.obo.versionextractor.XMLVersionExtractor;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvFileSource;
@@ -10,18 +12,18 @@ import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.Objects;
 
-public class TestOWLVersionExtractor {
+public class TestXMLVersionExtractor {
 
     @ParameterizedTest
-    @CsvFileSource(resources = "/owl-testcases.csv", numLinesToSkip = 1, delimiter = ',')
+    @CsvFileSource(resources = "/xml-testcases.csv", numLinesToSkip = 1, delimiter = ',')
     public void testExtractVersionIriFromXMLExtractsVersionInfoCorrectlyIfPresent(
             String file,
             String ontologyIri,
             String versionIri,
             String versionInfo) throws URISyntaxException, IOException {
-        final String s = Files.readString(Paths.get(Objects.requireNonNull(getClass().getResource("/owl-testcases/" + file)).toURI()));
+        final String s = Files.readString(Paths.get(Objects.requireNonNull(getClass().getResource("/xml-testcases/" + file)).toURI()));
         final Version version = new Version();
-        new FSVersionExtractor().extract(s, version);
+        new XMLVersionExtractor().extract(s, version);
         Assertions.assertEquals(ontologyIri, version.getOwlOntologyIri());
         Assertions.assertEquals(versionIri, version.getOwlVersionIri());
         Assertions.assertEquals(versionInfo, version.getOwlVersionInfo());
